Avoid instrumenting complex constructors
If there is branching logic before object initialization is complete, avoid instrumenting the entire constructor. Java doesn't seem to be able to generate these constructors (since the compiler forces the super constructor call to be the first statement of the constructor), but bytecode generated by other compilers/etc (e.g., Groovy) may give us hard-to-trace constructors due to the local entry on the stackmap frame switching from uninitialized-this to this after the super constructor call (so you can't simply wrap the entire thing in a giant try/catch like we do for everything else).

Since this is more of an edge case, we're just going to avoid instrumenting such constructors for now.